Split test_views.py into multiple modules
