+++ tardis/tardis_portal/tests/views/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tests for view methods in tardis/tardis_portal/views/
+"""
